fix(admin): scope the admin's CSS the same way in dev and release
The dev and release builds each carried their own copy of the scoper,
with a comment inviting them to diverge — and they had. Only the release
copy learned that a selector list must be split on top-level commas, that
a nested `&` variant is already scoped, and that `.dark` sits on the
wrapper rather than under it.

So the dev build still emitted `.nextly-admin .dark`, which cannot
match, mangled every dark variant, and passed `html`, `body`, `:root`
and `*` through untouched — the admin's reset landing on the host page,
which is the leak the release build already refused to ship.

One module now, because isolation has one right answer. The watcher
reports what escapes instead of failing, since a release build failing is
useful and a watcher dying mid-session is not.
