+++ packages/admin/scripts/build-css-fast.mjs
@@ -31,6 +31,8 @@ import { fileURLToPath } from "url";
 import postcss from "postcss";
 import tailwindcss from "@tailwindcss/postcss";
 
+import { findUnscopedRules, scopeCss } from "./lib/css-scope.mjs";
+
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 const rootDir = path.resolve(__dirname, "..");
 
@@ -41,92 +43,6 @@ const outputFile = path.join(outputDir, "globals.css");
 // Reuse the processor across calls — Tailwind v4's content-scan caches
 // live inside the plugin instance and amortise across rebuilds.
 const processor = postcss([tailwindcss()]);
-
-// .nextly-admin scoping — same logic as scripts/build-css.mjs. Duplicated
-// here on purpose so the dev and production build pipelines can evolve
-// independently if we ever need different isolation behaviour. If/when
-// they stabilise, this can be hoisted into a shared module.
-const ALREADY_SCOPED = [
-  /^\.nextly-admin/,
-  /^\.dark\s*\.nextly-admin/,
-  /^\.nextly-admin\.dark/,
-  /^:root/,
-  /^\*/,
-  /^html/,
-  /^body/,
-  /^@keyframes/,
-  /^@font-face/,
-  /^@media/,
-  /^@supports/,
-  /^@layer/,
-  /^@property/,
-];
-
-function scopeSelector(selector) {
-  selector = selector.trim();
-  for (const p of ALREADY_SCOPED) {
-    if (p.test(selector)) return selector;
-  }
-  if (selector.startsWith("@")) return selector;
-  if (selector === ":root") return ".nextly-admin";
-  if (selector === "*" || selector === "*::before" || selector === "*::after") {
-    return `.nextly-admin ${selector}`;
-  }
-  if (selector.includes(",")) {
-    return selector
-      .split(",")
-      .map(s => scopeSelector(s.trim()))
-      .join(", ");
-  }
-  return `.nextly-admin ${selector}`;
-}
-
-function scopeCss(css) {
-  const lines = css.split("\n");
-  const out = [];
-  let inKeyframes = false;
-  let braceCount = 0;
-
-  for (const line of lines) {
-    if (line.includes("@keyframes")) {
-      inKeyframes = true;
-      out.push(line);
-      continue;
-    }
-    if (inKeyframes) {
-      braceCount += (line.match(/{/g) || []).length;
-      braceCount -= (line.match(/}/g) || []).length;
-      if (braceCount <= 0) {
-        inKeyframes = false;
-        braceCount = 0;
-      }
-      out.push(line);
-      continue;
-    }
-    const trimmed = line.trim();
-    if (
-      trimmed.startsWith("@media") ||
-      trimmed.startsWith("@supports") ||
-      trimmed.startsWith("@layer") ||
-      trimmed.startsWith("@font-face") ||
-      trimmed.startsWith("@property")
-    ) {
-      out.push(line);
-      continue;
-    }
-    if (line.includes("{") && !trimmed.startsWith("@")) {
-      const [selector, ...rest] = line.split("{");
-      if (selector.includes(".nextly-admin") || selector.trim().startsWith("--")) {
-        out.push(line);
-        continue;
-      }
-      out.push(`${scopeSelector(selector)}{${rest.join("{")}`);
-    } else {
-      out.push(line);
-    }
-  }
-  return out.join("\n");
-}
 
 export async function buildCssFast() {
   const start = performance.now();
@@ -140,7 +56,11 @@ export async function buildCssFast() {
   await fs.writeFile(outputFile, scoped);
   const ms = Math.round(performance.now() - start);
   const sizeKB = (Buffer.byteLength(scoped) / 1024).toFixed(1);
-  return { ms, sizeKB, outputFile };
+  // Reported, not thrown: the production build fails on a leak, but killing a
+  // watcher mid-session would cost more than the warning is worth. Surfacing
+  // it here means a leak is seen when it is introduced rather than at release.
+  const unscoped = findUnscopedRules(scoped);
+  return { ms, sizeKB, outputFile, unscoped };
 }
 
 // Direct CLI invocation: `node scripts/build-css-fast.mjs`
